Backend/data/repo_temp/repopilot_wvzwcrgq/backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()

GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")

print("DEBUG GEMINI KEY:", GEMINI_API_KEY)

if not GEMINI_API_KEY:
    raise RuntimeError("GEMINI_API_KEY missing in backend/.env")


